# RUM-3796: Optimise `BatchFileOrchestator` performance