--- a/dd-sdk-android-core/src/main/kotlin/com/datadog/android/core/internal/persistence/file/FilePersistenceConfig.kt
+++ b/dd-sdk-android-core/src/main/kotlin/com/datadog/android/core/internal/persistence/file/FilePersistenceConfig.kt
@@ -25,6 +25,6 @@ internal data class FilePersistenceConfig(
         internal const val OLD_FILE_THRESHOLD: Long = 18L * 60L * 60L * 1000L // 18 hours
         internal const val MAX_DISK_SPACE: Long = 128 * MAX_BATCH_SIZE // 512 MB
         internal const val MAX_DELAY_BETWEEN_MESSAGES_MS = 5000L
-        internal const val CLEANUP_FREQUENCY_THRESHOLD_MS = 1000L // 1s
+        internal const val CLEANUP_FREQUENCY_THRESHOLD_MS = 5000L // 5s
     }
 }